Improve logging by prepending log lines with identifiers.
In order to better track what happens for a given container in Mesos logs
we need to correlate prepare and cleanup operation with some kind of
identifier. We take the identifiers provided with the container to
keep track of it.

NOTE: unfortunately the identifier is different for an isolator and
a hook because different information is provided in each case.