Add support for isolator watch callback (#15)
Add support for isolator watch callback

Caveat: in the current implementation, the watch command is limited by a time out, and called only once in the container lifetime (the very same way it is done in native by Mesos isolators). This should be changed to a model where the command is called at a given frequency (and still controlled by a timeout).