Define several instances of modules
Mesos does not accept to instanciate several instances of the same
module (likely because there is no generic module like ours).
Yet users (criteo included) might want to have several "isolators" or
"hooks" having different features.

Our workaround so far was to create a wrapper script that would call the
others but it is quite unpractical introduce coupling.

This patch creates a fixed number of isolator/hooks that can be used.
It is quite easy to increase their number.

Change-Id: I27e92ed62ca751d01a5b18f04760534bd3985cd4